% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdbtools_tools.R
\name{mdb_prop}
\alias{mdb_prop}
\title{Get Properties List From MDB Database}
\usage{
mdb_prop(path, name = NULL, propcol = "LvProp", version = FALSE)
}
\arguments{
\item{path}{Path to \code{.mdb}/\code{.accdb} file.}

\item{name}{Object name (\code{table}, \code{query}, etc.).}

\item{propcol}{Property column name. Defaults to \code{LvProp}.}

\item{version}{Logical; when \code{TRUE}, return mdbtools version.}
}
\value{
Character scalar containing the command output.
}
\description{
\code{mdb-prop} prints a properties list from an MDB database.
\code{name} is the name of the table, query, or other object.
\code{propcol} is the name of the \code{MSysObjects} column containing properties and
defaults to \code{LvProp}.
}
