% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdbtools_tools.R
\name{mdb_schema}
\alias{mdb_schema}
\title{Generate Schema Creation DDL}
\usage{
mdb_schema(
  path,
  table = NULL,
  namespace = NULL,
  backend = c("access", "sybase", "oracle", "postgres", "mysql", "sqlite"),
  drop_table = FALSE,
  not_null = TRUE,
  default_values = FALSE,
  not_empty = FALSE,
  comments = TRUE,
  indexes = TRUE,
  relations = TRUE
)
}
\arguments{
\item{path}{Path to \code{.mdb}/\code{.accdb} file.}

\item{table}{Single table option, equivalent to \code{-T/--table}. Default is to
export all user tables.}

\item{namespace}{Prefix identifiers with namespace, equivalent to
\code{-N/--namespace}.}

\item{backend}{Target DDL dialect. Supported values are \code{access}, \code{sybase},
\code{oracle}, \code{postgres}, \code{mysql}, and \code{sqlite}.}

\item{drop_table}{Issue \verb{DROP TABLE} statements.}

\item{not_null}{Include \verb{NOT NULL} constraints.}

\item{default_values}{Include \code{DEFAULT} values.}

\item{not_empty}{Include \verb{CHECK <> ''} constraints.}

\item{comments}{Include \verb{COMMENT ON} statements.}

\item{indexes}{Export indexes.}

\item{relations}{Export foreign key constraints.}
}
\value{
Character scalar containing the generated DDL.
}
\description{
\code{mdb-schema} is a utility program distributed with MDB Tools.
It produces DDL (data definition language) output for the given database.
This can be passed to another database to create a replica of the original
Access table format.
}
