\name{mdbtoolr-package}
\alias{mdbtoolr-package}
\title{mdbtoolr: DBI Interface to mdbtools}
\description{
The \pkg{mdbtoolr} package provides a \pkg{DBI} backend that reads
Microsoft Access \code{.mdb} and \code{.accdb} files through
external \code{mdbtools} commands.
}
\seealso{
\code{\link{Mdb}}
}
\keyword{package}
